Working api

+++ Project_14_MERN API/backend/controllers/goalController.js
@@ -2,13 +2,16 @@ const asyncHandler = require('express-async-handler')
 
 // Bring in model
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
+
 
 
 // @desc Get goals
 // @route Get /api/goals
 // @access Private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    // Get user goal
+    const goals = await Goal.find({ user: req.user.id})
 
 
     res.status(200).json(goals)
@@ -23,7 +26,8 @@ const setGoal = asyncHandler(async (req, res) => {
         throw new Error('Please add a text field')
     }
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
 
     res.status(200).json(goal)
@@ -39,6 +43,23 @@ const updateGoal = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error('Goal not found')
     }
+
+    // Get user
+    const user = await User.findById(req.user.id)
+    // Check for user
+    if (!user){
+        // Not authorized
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Check that the logged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+        // Not authorized
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true})
 
 
@@ -55,9 +76,26 @@ const deleteGoal = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error('Goal not found')
     }
+
+    // Get user
+    const user = await User.findById(req.user.id)
+    // Check for user
+    if (!user){
+        // Not authorized
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Check that the logged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+        // Not authorized
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
     const deleteGoal = await Goal.findByIdAndDelete(req.params.id)
 
-    res.status(200).json({message: 'Delete goal' + req.params.id})
+    res.status(200).json({message: 'Delete goal: ' + req.params.id})
 })
 module.exports = {
     getGoals, setGoal, updateGoal, deleteGoal
